Use CGSizeFromGADAdSize to size our banners
The GADSize given to us may contain a zeroed CGSize when using smart
banners. Using the CGSize in the GADAdSize should not be done according
to the docs, instead this function is used to obtain a usable CGSize.

Also includes up to date google ad mob headers
and a universal build target.